requirements.txt
streamlit
pandas
Pillow
folium
matplotlib
streamlit_folium
plotly
# bamboolib
